+++ include/boost/multiprecision/traits/max_digits10.hpp
@@ -39,7 +39,7 @@ struct calc_max_digits10_s
       // never be exactly an integer so we can replace by trunc(log10(2) * d) + 2
       // and avoid the call to ceil:
       //
-      return static_cast<std::size_t>(0.301029995663981195213738894724493026768189881462108541310 * d) + 2;
+      return static_cast<std::size_t>(static_cast<std::size_t>(0.301029995663981195213738894724493026768189881462108541310 * static_cast<double>(d)) + 2u);
    }
    static constexpr const std::size_t value = max_digits_10(digits);
 };
@@ -54,7 +54,7 @@ struct calc_digits10
       // https://www.exploringbinary.com/number-of-digits-required-for-round-trip-conversions/
       // and references therein.
       //
-      return static_cast<unsigned>(0.301029995663981195213738894724493026768189881462108541310 * (d - 1));
+      return static_cast<unsigned>(0.301029995663981195213738894724493026768189881462108541310 * static_cast<double>(d - 1u));
    }
    static constexpr const unsigned value = digits_10(digits);
 };
@@ -69,7 +69,7 @@ struct calc_digits10_s
       // https://www.exploringbinary.com/number-of-digits-required-for-round-trip-conversions/
       // and references therein.
       //
-      return static_cast<std::size_t>(0.301029995663981195213738894724493026768189881462108541310 * (d - 1));
+      return static_cast<std::size_t>(0.301029995663981195213738894724493026768189881462108541310 * static_cast<double>(d - 1u));
    }
    static constexpr const std::size_t value = digits_10(digits);
 };
